fix: 合併後才出現的 10 條 lint 歸零 (F170-13, F170-14)
sprint branch 合併後 `task check` 掛在 lint,測試全綠。其中 goconst 兩條是
A/B/C 三份合併後 package 層計數才被推過門檻的跨檔交互,單一隔離樹重現不了。

F170-13 gosec G115 ×3:listPageBounds 改回 int32。原本三個呼叫點各寫
int32(limit+1),是真的 narrowing —— 舊的 int 簽章沒有任何地方保證值有界,只有
函式本體有。改回傳型別把「不可能溢位」從註解變成型別保證,轉換全部消失。
新增 TestF170_13_ListPageBoundsClampsWithinInt32 直接釘 clamp(handler 測試
釘不住:它們塞 500 筆,clamp 放寬到 300 也一樣綠,因為 fixture 先用完)。

F170-14 結構清理,行為不變:
- gocyclo classify(19)/classifyCallResult(20) → 抽 unwrapSingleOperand /
  isTypeExpression / classifyExprs / classifyCallee / classifyPackageLocalCall。
  求值順序逐條保留;isTypeExpression 刻意不含 StarExpr(那是 pass-through
  operand,混進去會把指標解參考判成 static)。
- SA1019 → 保留 parser.ParseDir + 具名 nolint。實測:ParseDir 解析 43 檔,
  非測試 .go 也是 43,雙向差集空;build-tagged 檔有 0 個是非測試檔,所以
  build constraint 在這個掃描面上沒有東西可排除。
- goconst → newToolResultErrorFn / goStringTypeName 兩個具名常數;"true" 那條
  nolint 並寫明四個字面值是一組 truthy-env 值、抽單一常數反而更難讀。
- QF1011 → 刪冗餘的 var _ []db.PendingProposal。

驗收:F170-08 閘門判定結果零變化 —— SITE_COUNT 266 → 266,剝掉行號後的
(file, verdict, kind) multiset 逐字相同,紅的清單前後皆空。

+++ internal/mcp/tools_gtd.go
@@ -627,15 +627,23 @@ const (
 // two other parallel work-streams are copying from right now, and rewriting it
 // underneath them would break all three at merge. Folding it in is a
 // follow-up, not a silent drive-by.
-func listPageBounds(rawLimit, rawOffset int32) (limit, offset int) {
-	limit = int(rawLimit)
+// [F170-13] Returns int32, the width the store's paging API takes, so the
+// three call sites hand the result straight through with no conversion. The
+// earlier version returned int and each caller wrote int32(limit+1) — three
+// narrowing conversions that gosec G115 flags as potential overflow, and
+// correctly so: nothing in the TYPE said the value was bounded, only this
+// function's body did. Returning int32 moves "cannot overflow" from a comment
+// into the signature. limit is clamped to listPageMaxLimit here, so limit+1 at
+// a call site is at most 201 and cannot wrap.
+func listPageBounds(rawLimit, rawOffset int32) (limit, offset int32) {
+	limit = rawLimit
 	if limit <= 0 {
 		limit = listPageDefaultLimit
 	}
 	if limit > listPageMaxLimit {
 		limit = listPageMaxLimit
 	}
-	offset = int(rawOffset)
+	offset = rawOffset
 	if offset < 0 {
 		offset = 0
 	}
@@ -656,11 +664,11 @@ func (s *Server) handleListProjects(ctx context.Context, args ListProjectsArgs) 
 
 	// limit+1 detects has_more without a second COUNT query — same trick as
 	// handleListTasks.
-	projects, err := s.gtd.ActiveProjectsPage(ctx, int32(limit+1), int32(offset))
+	projects, err := s.gtd.ActiveProjectsPage(ctx, limit+1, offset)
 	if err != nil {
 		return storeErrorResult("loading projects", err), nil
 	}
-	hasMore := len(projects) > limit
+	hasMore := len(projects) > int(limit)
 	if hasMore {
 		projects = projects[:limit]
 	}
@@ -1177,11 +1185,11 @@ func (s *Server) applyArtifactSideEffects(ctx context.Context, id uuid.UUID, art
 func (s *Server) handleListGoals(ctx context.Context, args ListGoalsArgs) (*mcp.CallToolResult, error) {
 	limit, offset := listPageBounds(args.Limit, args.Offset)
 
-	goals, err := s.gtd.ActiveGoalsPage(ctx, int32(limit+1), int32(offset))
+	goals, err := s.gtd.ActiveGoalsPage(ctx, limit+1, offset)
 	if err != nil {
 		return storeErrorResult("loading goals", err), nil
 	}
-	hasMore := len(goals) > limit
+	hasMore := len(goals) > int(limit)
 	if hasMore {
 		goals = goals[:limit]
 	}
